// Foodoz_Admin/foodoz-angular/src/app/@core/models/vendor/vendor.error.ts
export class VendorError {
    name_translations: Array<string>;
    tagline_translations: Array<string>;
    details_translations: Array<string>;
    meta: Array<string>;
    minimum_order: Array<string>;
    delivery_fee: Array<string>;
    area: Array<string>;
    address: Array<string>;
    longitude: Array<string>;
    latitude: Array<string>;
    image: Array<string>;
    user_id: Array<string>;
    categories: Array<string>;
    is_verified: Array<string>;

    // user related
    email: Array<string>;
    mobile_number: Array<string>;
    password: Array<string>;
}